if you import a file which is alread in workpath, no delete and importing again anymore

+++ scripts/P8Modules/ReadTable_Gui.py
@@ -39,8 +39,9 @@ class ReadTable_Gui(QtGui.QDialog):
 		self.ui.le_r.setText(QFileInfo(filename).fileName())
 		settings.setValue("dataPath",finfo.absolutePath())
 		if(os.path.exists(workpath + finfo.fileName())):
-			os.remove(str(workpath) + str(finfo.fileName()))
-			copyfile(filename, str(workpath) + str(finfo.fileName()))
+			if(str(finfo.absolutePath() + "/") != str(workpath)):
+				os.remove(str(workpath) + str(finfo.fileName()))
+				copyfile(filename, str(workpath) + str(finfo.fileName()))
 		else:
 			copyfile(filename,str(workpath) + str(finfo.fileName()))
 		self.loadTable(str(workpath) + str(finfo.fileName()))
